+++ nvim/lua/plugins/blink.lua
@@ -1,0 +1,29 @@
+return {
+  "saghen/blink.cmp",
+  version = "1.*",
+  dependencies = {
+    "rafamadriz/friendly-snippets",
+    "L3MON4D3/LuaSnip",
+  },
+  ---@module 'blink.cmp'
+  ---@type blink.cmp.Config
+  opts = {
+    keymap = { preset = "super-tab" },
+    appearance = {
+      nerd_font_variant = "mono",
+    },
+    completion = {
+      documentation = { auto_show = true },
+    },
+    sources = {
+      default = { "lsp", "path", "snippets", "buffer" },
+    },
+    snippets = {
+      preset = "luasnip",
+    },
+    fuzzy = {
+      implementation = "prefer_rust",
+    },
+  },
+  opts_extend = { "sources.default" },
+}